requirements.txt
fastapi
uvicorn
pymilvus
python-dotenv
pandas